--- Entidade.cpp
@@ -1,32 +1,0 @@
-#include "Entidade.h"
-
-Entidade::Entidade(int id,const char* png):
-	Ente(id,png)
-{
-	velocidadeY = 0;
-	chao = 0;
-	noChao = 0;
-	x = 60;
-	y = 60;
-}
-
-Entidade::~Entidade() {
-}
-
-void Entidade::setCoordenadas(float x1, float y1) {
-	x = x1;
-	y = y1;
-}
-
-void Entidade::setCorpo(float largura, float altura) {
-	corpo.setSize(Vector2f(largura, altura));
-	corpo.setTexture(&sprite);
-	corpo.setPosition(x, y);
-}
-
-void Entidade::setSoCorpo(float largura, float altura) {
-	corpo.setSize(Vector2f(largura, altura));
-}
-
-
-const float Entidade::gravidade = 9.8f;
