+++ Gerenciadores/Gerenciador_Colisoes.h
@@ -4,8 +4,8 @@
 #include "../Obstaculos/Plataforma.h"
 #include "../Personagens/Inimigo.h"
 #include "../Personagens/Heroi.h"
-#include "../Entidade.h"
-#include "../Projetil.h"
+#include "../Entidades/Entidade.h"
+#include "../Entidades/Projetil.h"
 
 using namespace std;
 
@@ -13,20 +13,20 @@ namespace Gerenciadores {
 	class Gerenciador_Colisoes
 	{
 	private:
-		list<Obstaculos::Obstaculo*> LObst;
-		list<Personagens::Inimigo*> LIni;
-		list<Projetil*> LProjetil;
-		Personagens::Heroi *jog1;
-		Personagens::Heroi *jog2;
+		list<Entidades::Obstaculos::Obstaculo*> LObst;
+		list<Entidades::Personagens::Inimigo*> LIni;
+		list<Entidades::Projetil*> LProjetil;
+		Entidades::Personagens::Heroi *jog1;
+		Entidades::Personagens::Heroi *jog2;
 	public:
 		Gerenciador_Colisoes();
 		~Gerenciador_Colisoes();
 		void trocaFase();
-		void setJogadores(Personagens::Heroi* j1, Personagens::Heroi* j2);
-		Personagens::Heroi* getJogador2() { return jog2; }
-		void includeEntidade(Entidade* ent);
-		void removeEntidade(Entidade* ent);
-		int veriColisao(Entidade* ent,Entidade* sl);
+		void setJogadores(Entidades::Personagens::Heroi* j1, Entidades::Personagens::Heroi* j2);
+		Entidades::Personagens::Heroi* getJogador2() { return jog2; }
+		void includeEntidade(Entidades::Entidade* ent);
+		void removeEntidade(Entidades::Entidade* ent);
+		int veriColisao(Entidades::Entidade* ent,Entidades::Entidade* sl);
 		void verificaObs();
 		void verificaIni();
 		void veriProjetil();
