+++ Jogo.vcxproj
@@ -148,21 +148,20 @@ sfml-system-d.lib;%(AdditionalDependencies)</AdditionalDependencies>
     <ClCompile Include="Menu.cpp" />
     <ClCompile Include="Personagens\Samurai.cpp" />
     <ClCompile Include="Ente.cpp" />
-    <ClCompile Include="Entidade.cpp" />
+    <ClCompile Include="Entidades\Entidade.cpp" />
     <ClCompile Include="Obstaculos\Espinho.cpp" />
     <ClCompile Include="Fases\Fase.cpp" />
     <ClCompile Include="Gerenciadores\Gerenciador_Colisoes.cpp" />
     <ClCompile Include="Gerenciadores\Gerenciado_Grafico.cpp" />
     <ClCompile Include="Personagens\Yokai.cpp" />
-    <ClCompile Include="Personagens\Efeitos.cpp" />
     <ClCompile Include="Personagens\Inimigo.cpp" />
-    <ClCompile Include="Jogo.cpp" />
+    <ClCompile Include="Yokai_Slayer.cpp" />
     <ClCompile Include="Listas\ListaEntidades.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="Obstaculos\Obstaculo.cpp" />
     <ClCompile Include="Personagens\Personagem.cpp" />
     <ClCompile Include="Obstaculos\Plataforma.cpp" />
-    <ClCompile Include="Projetil.cpp" />
+    <ClCompile Include="Entidades\Projetil.cpp" />
     <ClCompile Include="Personagens\Lobo.cpp" />
     <ClCompile Include="Personagens\Heroi.cpp" />
     <ClCompile Include="Obstaculos\Rio.cpp" />
@@ -174,21 +173,20 @@ sfml-system-d.lib;%(AdditionalDependencies)</AdditionalDependencies>
     <ClInclude Include="Menu.h" />
     <ClInclude Include="Personagens\Samurai.h" />
     <ClInclude Include="Ente.h" />
-    <ClInclude Include="Entidade.h" />
+    <ClInclude Include="Entidades\Entidade.h" />
     <ClInclude Include="Obstaculos\Espinho.h" />
     <ClInclude Include="Fases\Fase.h" />
     <ClInclude Include="Gerenciadores\Gerenciador_Colisoes.h" />
     <ClInclude Include="Gerenciadores\Gerenciado_Grafico.h" />
     <ClInclude Include="Personagens\Yokai.h" />
-    <ClInclude Include="Personagens\Efeitos.h" />
     <ClInclude Include="Personagens\Inimigo.h" />
-    <ClInclude Include="Jogo.h" />
+    <ClInclude Include="Yokai_Slayer.h" />
     <ClInclude Include="Listas\Lista.h" />
     <ClInclude Include="Listas\ListaEntidades.h" />
     <ClInclude Include="Obstaculos\Obstaculo.h" />
     <ClInclude Include="Personagens\Personagem.h" />
     <ClInclude Include="Obstaculos\Plataforma.h" />
-    <ClInclude Include="Projetil.h" />
+    <ClInclude Include="Entidades\Projetil.h" />
     <ClInclude Include="Personagens\Lobo.h" />
     <ClInclude Include="Personagens\Heroi.h" />
     <ClInclude Include="Obstaculos\Rio.h" />
