+++ Listas/ListaEntidades.cpp
@@ -2,16 +2,15 @@
 #include "../Personagens/Heroi.h"
 #include "../Personagens/Yokai.h"
 
-using namespace Personagens;
 namespace Listas {
 
 	class Obstaculo;
 
 	/*===== construtora =====*/
 	ListaEntidade::ListaEntidade() :
-		listaEntidades(new Lista<Entidade*>()),
-		chefoes(new std::vector<Yokai*>()),
-		plataforma_Yokai(new std::queue<Entidade*>())
+		listaEntidades(new Lista<Entidades::Entidade*>()),
+		chefoes(new std::vector<Entidades::Personagens::Yokai*>()),
+		plataforma_Yokai(new std::queue<Entidades::Entidade*>())
 	{
 		pos_Yokai = 0;
 		zona_Yokai = false;
@@ -41,18 +40,18 @@ namespace Listas {
 		gc->executar();//executar do gerenciador de colisoes 
 
 
-		for (Lista<Entidade*>::iterator it = listaEntidades->begin(); it != listaEntidades->end(); it++)
+		for (Lista<Entidades::Entidade*>::iterator it = listaEntidades->begin(); it != listaEntidades->end(); it++)
 		{
 			if ((*it)->getId() == 3) {//se for uma plataforma
 				(*it)->executar();
 			}
 			if ((*it)->getId() == 5) {//se for um projetil
-				if (static_cast<Projetil*>(*it)->getApareceu() == true) {//se ele ja apareceu
+				if (static_cast<Entidades::Projetil*>(*it)->getApareceu() == true) {//se ele ja apareceu
 						(*it)->executar();//executa o projetil
 				}
 			}
 			else if ((*it)->getId() == 6) {//se for um Yokai
-				if (static_cast<Personagens::Yokai*>(*it)->getAtivo() == true) {
+				if (static_cast<Entidades::Personagens::Yokai*>(*it)->getAtivo() == true) {
 					(*it)->executar();
 				}
 			}
@@ -63,20 +62,20 @@ namespace Listas {
 				(*it)->executar();
 
 				if ((*it)->getId() == 1) {//se for um jogador
-					encontraZonaYokai(static_cast<Heroi*>(*it));//verifico qual zona ele tá
+					encontraZonaYokai(static_cast<Entidades::Personagens::Heroi*>(*it));//verifico qual zona ele tá
 				}
 			}
 		}
 	}
 
 	/*===== inclui uma entidade na lista e verifica se eh Yokai ou projetil =====*/
-	void ListaEntidade::Incluir(Entidade* entidade, Gerenciadores::Gerenciador_Colisoes* gc) {
+	void ListaEntidade::Incluir(Entidades::Entidade* entidade, Gerenciadores::Gerenciador_Colisoes* gc) {
 
 		listaEntidades->adicionarElemento(entidade);//adiciono na lista
 		gc->includeEntidade(entidade);//incluo no gerenciador de colisoes
 
 		if (entidade->getId() == 6) {//se for Yokai adiciono-o no vector
-			chefoes->push_back(static_cast<Yokai*>(entidade));//coloco o Yokai no vetor de Yokai
+			chefoes->push_back(static_cast<Entidades::Personagens::Yokai*>(entidade));//coloco o Yokai no vetor de Yokai
 			
 		}
 
@@ -88,10 +87,10 @@ namespace Listas {
 
 			tam = (int)chefoes->size();
 			while (inseri != true && i<tam) {
-				Yokai* Yokai = chefoes->operator[](i);
+				Entidades::Personagens::Yokai* Yokai = chefoes->operator[](i);
 
 				if (Yokai->getNum_Projetil() < Yokai->getMaxProjetil()) {
-					Yokai->criaProjeteis(static_cast<Projetil*>(entidade));
+					Yokai->criaProjeteis(static_cast<Entidades::Projetil*>(entidade));
 					inseri = true;
 				}
 				else {
@@ -106,17 +105,17 @@ namespace Listas {
 	}
 
 	/*=== Inclui entidade na lista ===*/ //*************** nao está sendo usado!!
-	void ListaEntidade::Include(Entidade* entidade) {
+	void ListaEntidade::Include(Entidades::Entidade* entidade) {
 		listaEntidades->adicionarElemento(entidade);
 	}
 
 	/*=== incluir para o salvamento!! ===*/
-	void ListaEntidade::IncluirSalvamento(Entidade* entidade, Gerenciadores::Gerenciador_Colisoes* gc) {
+	void ListaEntidade::IncluirSalvamento(Entidades::Entidade* entidade, Gerenciadores::Gerenciador_Colisoes* gc) {
 		listaEntidades->adicionarElemento(entidade);//adiciono na lista
 		gc->includeEntidade(entidade);//incluo no gerenciador de colisoes
 
 		if (entidade->getId() == 6) {//se for Yokai adiciono-o no vector
-			chefoes->push_back(static_cast<Yokai*>(entidade));//coloco o Yokai no vetor de Yokai
+			chefoes->push_back(static_cast<Entidades::Personagens::Yokai*>(entidade));//coloco o Yokai no vetor de Yokai
 
 		}
 		if (entidade->getId() == 5) {//se for projetil, insiro no Yokai que tiver espaço livre
@@ -126,10 +125,10 @@ namespace Listas {
 			int tam;
 			tam =(int) chefoes->size();
 			while (inseri != true && i < tam) {
-				Yokai* Yokai = chefoes->operator[](i);
+				Entidades::Personagens::Yokai* Yokai = chefoes->operator[](i);
 
 				if (Yokai->getNum_Proj_Salv() < Yokai->getNum_Projetil()) {
-					Yokai->criaProjeteisSalv((static_cast<Projetil*>(entidade)));
+					Yokai->criaProjeteisSalv((static_cast<Entidades::Projetil*>(entidade)));
 					inseri = true;
 				}
 				else {
@@ -147,7 +146,7 @@ namespace Listas {
 	/*===== seta o gerenciador gráfico em todas as instâncias da lista =====*/
 	void ListaEntidade::setGG(Gerenciadores::Gerenciado_Grafico* gg) {
 
-		for (Lista<Entidade*>::iterator it = listaEntidades->begin(); it != listaEntidades->end(); it++)
+		for (Lista<Entidades::Entidade*>::iterator it = listaEntidades->begin(); it != listaEntidades->end(); it++)
 		{
 			(*it)->setGerenciador(gg);
 		}
@@ -159,16 +158,16 @@ namespace Listas {
 		bool temMortos = false;
 
 
-		for (Lista<Entidade*>::iterator it = listaEntidades->begin(); it != listaEntidades->end() && temMortos == false; it++) {
+		for (Lista<Entidades::Entidade*>::iterator it = listaEntidades->begin(); it != listaEntidades->end() && temMortos == false; it++) {
 			tam++;
 			if ((*it)->getId() == 4 || (*it)->getId() == 6 || (*it)->getId() == 7) {
-				if (static_cast<Inimigo*>(*it)->getVidas() <= 0) {
+				if (static_cast<Entidades::Personagens::Inimigo*>(*it)->getVidas() <= 0) {
 					temMortos = true;
 					return listaEntidades->posicao(*it);
 				}
 			}
 			if ((*it)->getId() == 5) {
-				if (static_cast<Projetil*>(*it)->getNoChao() == true || static_cast<Projetil*>(*it)->getApagado()==true) {
+				if (static_cast<Entidades::Projetil*>(*it)->getNoChao() == true || static_cast<Entidades::Projetil*>(*it)->getApagado()==true) {
 					temMortos = true;
 					return listaEntidades->posicao(*it);
 				}
@@ -177,18 +176,15 @@ namespace Listas {
 		return (-1);//se nao houver mortos, retorna -1
 	}
 
-	/*===== remove uma determinada entidade da lista =====*/
-	void ListaEntidade::MatarEntidade(Entidade* ent, Gerenciadores::Gerenciador_Colisoes* gc) {
-
+	void ListaEntidade:: MatarEntidade(Entidades::Entidade* ent, Gerenciadores::Gerenciador_Colisoes* gc) {
 		listaEntidades->removerElemento(ent);
 		gc->removeEntidade(ent);
-
-		
 	}
+
 
 	/*===== remove uma entidade com uma determinada posição na lista =====*/
 	void ListaEntidade::matarEntidadePos(int pos, Gerenciadores::Gerenciador_Colisoes * gc){
-		Entidade* ent = listaEntidades->operator[](pos);
+		Entidades::Entidade* ent = listaEntidades->operator[](pos);
 
 		if (ent->getId() == 7 || ent->getId() == 4) {
 			MatarEntidade(ent, gc);
@@ -207,7 +203,7 @@ namespace Listas {
 
 
 	/*===== seta se um jogador está na zona de um determinado Yokai... =====*/
-	void ListaEntidade::encontraZonaYokai(Heroi* jog) {
+	void ListaEntidade::encontraZonaYokai(Entidades::Personagens::Heroi* jog) {
 		bool achou = false;
 
 		for (int i = 0; i < chefoes->size() && achou != true; i++) {
@@ -216,13 +212,13 @@ namespace Listas {
 		}
 	}
 
-	vector<Entidade*> ListaEntidade::returnVec() {
-		vector<Entidade*> aux;
+	vector<Entidades::Entidade*> ListaEntidade::returnVec() {
+		vector<Entidades::Entidade*> aux;
 
-		for (Lista<Entidade*>::iterator it = listaEntidades->begin(); it != listaEntidades->end(); it++) {
+		for (Lista<Entidades::Entidade*>::iterator it = listaEntidades->begin(); it != listaEntidades->end(); it++) {
 			if ((*it)->getId() == 1 || (*it)->getId() == 2 || (*it)->getId() == 4 ||
 				(*it)->getId() == 7 || (*it)->getId() == 6) {
-				Personagem* p = static_cast<Personagem*>(*it);
+				Entidades::Personagens::Personagem* p = static_cast<Entidades::Personagens::Personagem*>(*it);
 				if (p->verificaVida()) {
 					aux.push_back(*it);
 				}
